adding unit test for reusing vars
